Enhance PoolInfoTab and TechnicianReportPage with payment adjustments and validation
- Updated PoolInfoTab to provide a clearer placeholder for payment unit input and added minimum and step properties for better user guidance.
- Enhanced TechnicianReportPage by introducing optional fields for extras and discounts, including validation for their amounts and descriptions.
- Refactored report generation logic to accommodate new adjustment parameters, improving the overall functionality and user experience in report handling.

+++ src/schemas/defaultSchemas.ts
@@ -42,7 +42,7 @@ export const defaultSchemas = {
     invalid_type_error: "Language must be 'English', 'Portuguese' or 'Spanish'."
   }),
   monthlyPayment: z.number().nullable().optional(),
-  paymentUnit: z.coerce.number().int().min(1, { message: 'Payment unit must be 1 or higher.' }).optional(),
+  paymentUnit: z.coerce.number().min(1, { message: 'Payment unit must be 1 or higher.' }).optional(),
   volumeInGallons: z.preprocess(
     (v) => (v === '' || v === null || v === undefined ? undefined : v),
     z.coerce.number().int().min(1, { message: 'Volume must be greater than 0.' }).optional()
